Add forgotten err assignment
Looking at the error check below, it looks like we wanted to handle
the error coming from DeleteMulti, but forgot to assign the var.

Please review this thoroughly. Should we indeed want to ignore this
error, we should remove the error check below as err can't ever be
non-nil here.